fittrack-backend/requirements.txt
# fittrack-backend/requirements.txt
# 실제 사용하는 패키지만 포함

# Web Framework
fastapi
uvicorn[standard]
gunicorn

# HTTP Client
requests

# Database
sqlalchemy
alembic
psycopg2-binary

# Data Validation & Settings
pydantic
pydantic-settings

# Authentication
python-jose[cryptography]
bcrypt==3.2.0
python-multipart

# Environment Variables
python-dotenv

# AI (Gemini)
google-generativeai

# Excel Export (매출/지출 엑셀 다운로드용)
pandas
openpyxl
